hello

+++ ft_chek_form.c
@@ -6,7 +6,7 @@
 /*   By: zalaksya <zalaksya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/25 10:17:04 by zalaksya          #+#    #+#             */
-/*   Updated: 2024/11/27 09:50:31 by zalaksya         ###   ########.fr       */
+/*   Updated: 2024/11/28 11:45:28 by zalaksya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int	ft_chek_form(int i, va_list p)
 	else if (i == 's')
 		count += ft_putstr(va_arg(p, char *));
 	else if (i == 'd' || i == 'i')
-		count += ft_putnbr(va_arg(p, int));
+		count += ft_putnbr(va_arg(p, long));
 	else if (i == 'x' || i == 'X')
 		count += ft_puthex(va_arg(p, unsigned int), i);
 	else if (i == 'u')
