+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: zalaksya <zalaksya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/24 11:09:33 by zalaksya          #+#    #+#             */
-/*   Updated: 2024/11/27 09:50:38 by zalaksya         ###   ########.fr       */
+/*   Updated: 2024/11/28 11:45:15 by zalaksya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 
 int	ft_putchar(char c);
 int	ft_putstr(char *str);
-int	ft_putnbr(int n);
+int	ft_putnbr(long n);
 int	ft_unsigned_putnbr(unsigned int n);
 int	ft_puthex(unsigned long n, int format);
 int	ft_chek_form(int i, va_list p);
